+++ wis2box-management/docker/wis2box.cron
@@ -1,4 +1,3 @@
 0 0 * * * su wis2box -c "wis2box data clean --days=$WIS2BOX_DATA_RETENTION_DAYS" > /proc/1/fd/1 2>/proc/1/fd/2
 0 1 * * * su wis2box -c "wis2box data archive" > /proc/1/fd/1 2>/proc/1/fd/2
-*/10 * * * * su wis2box -c "wis2box metadata station publish-collection" > /proc/1/fd/1 2>/proc/1/fd/2
 */10 * * * * su wis2box -c "echo 'wis2box.cron is alive'" > /proc/1/fd/1 2>/proc/1/fd/2
