+++ wis2box-management/wis2box/data/csv2bufr.py
@@ -19,21 +19,18 @@
 #
 ###############################################################################
 
-import json
 import logging
 from pathlib import Path
 from typing import Union
 
-from csv2bufr import MAPPINGS, transform as transform_csv
-
 from wis2box.data.base import BaseAbstractData
-from wis2box.metadata.station import get_valid_wsi
+from wis2box.api import execute_process
 
 LOGGER = logging.getLogger(__name__)
 
 
 class ObservationDataCSV2BUFR(BaseAbstractData):
-    """Observation data"""
+    """Synoptic observation data"""
     def __init__(self, defs: dict) -> None:
         """
         ObservationDataCSV2BUFR data initializer
@@ -47,16 +44,15 @@ class ObservationDataCSV2BUFR(BaseAbstractData):
 
         self.mappings = {}
 
-        LOGGER.debug(f'Loading template {self.template}')
-        if self.template.startswith('/'):
-            mapping_bufr4 = Path(self.template)
-        else:
-            mapping_bufr4 = Path(MAPPINGS) / self.template
+    def publish(self) -> bool:
+        """
+        Publish data
 
-        with mapping_bufr4.open() as fh1:
-            self.mappings['bufr4'] = json.load(fh1)
+        :returns: `bool` of result
+        """
 
-        self.station_metadata = None
+        LOGGER.debug('skip publish for call_synop_publish-plugin')
+        return True
 
     def transform(self, input_data: Union[Path, bytes],
                   filename: str = '') -> bool:
@@ -67,46 +63,24 @@ class ObservationDataCSV2BUFR(BaseAbstractData):
             LOGGER.debug('input_data is a Path')
             filename = input_data.name
 
-        if self.validate_filename_pattern(filename) is None:
+        file_match = self.validate_filename_pattern(filename)
+
+        if file_match is None:
             msg = f'Invalid filename format: {filename} ({self.file_filter})'
             LOGGER.error(msg)
             raise ValueError(msg)
 
-        LOGGER.debug('Generating BUFR4')
         input_bytes = self.as_bytes(input_data)
 
-        LOGGER.debug('Transforming data')
-        results = transform_csv(input_bytes.decode(),
-                                self.mappings['bufr4'])
-
-        LOGGER.debug('Iterating over BUFR messages')
-        for item in results:
-            wsi = item['_meta']['properties']['wigos_station_identifier']
-            if 'result' in item['_meta']:
-                if item['_meta']['result']['code'] != 1:
-                    msg = item['_meta']['result']['message']
-                    LOGGER.error(f'Transform returned {msg} for wsi={wsi}')
-                    self.publish_failure_message(
-                        description='csv2bufr transform error',
-                        wsi=wsi)
-                    continue
-            if get_valid_wsi(wsi) is None:
-                msg = f'Station {wsi} not in station list; skipping'
-                LOGGER.error(msg)
-                self.publish_failure_message(
-                        description='Station not in station list',
-                        wsi=wsi)
-                continue
-            LOGGER.debug('Setting obs date for filepath creation')
-            identifier = item['_meta']['id']
-            data_date = item['_meta']['properties']['datetime']
-
-            self.output_data[identifier] = item
-            self.output_data[identifier]['_meta']['relative_filepath'] = \
-                self.get_local_filepath(data_date)
-
+        process_name = 'wis2box-csv-process'
+        # execute process
+        inputs = {
+            "data": input_bytes.decode(),
+            "mapping": self.mappings['bufr4'],
+            "channel": self.topic_hierarchy,
+            "notify": self.enable_notification
+        }
+        result = execute_process(process_name, inputs)
+        for error in result['errors']:
+            LOGGER.error(f'{process_name} error: {error}')
         return True
-
-    def get_local_filepath(self, date_):
-        yyyymmdd = date_.strftime('%Y-%m-%d')
-        return (Path(yyyymmdd) / 'wis' / self.topic_hierarchy.dirpath)
